+++ src/main/java/me/MrRafter/framedupe/utils/ShulkerUtil.java
@@ -1,15 +1,11 @@
 package me.MrRafter.framedupe.utils;
 
-import org.bukkit.Material;
 import org.bukkit.block.ShulkerBox;
 import org.bukkit.inventory.Inventory;
 import org.bukkit.inventory.ItemStack;
 import org.bukkit.inventory.meta.BlockStateMeta;
-import org.bukkit.inventory.meta.BundleMeta;
 
-import java.util.List;
-
-public class ItemUtil {
+public class ShulkerUtil {
 
     /*
     * Metas as well as instanceof checks are slower than Material checks, but since we don't have to trigger this check
@@ -27,22 +23,5 @@ public class ItemUtil {
     * */
     public static Inventory getShulkerInventory(ItemStack shulkerItem) {
         return ((ShulkerBox) ((BlockStateMeta) shulkerItem.getItemMeta()).getBlockState()).getInventory();
-    }
-
-    /*
-     * We can check Material here since, unlike with shulkers, it hasn't been changed between versions yet.
-     */
-    public static boolean isNonEmptyBundle(ItemStack itemStack) {
-        if (itemStack == null) return false;
-        if (!itemStack.getType().equals(Material.BUNDLE)) return false;
-        if (!itemStack.hasItemMeta()) return false;
-        return ((BundleMeta) itemStack.getItemMeta()).hasItems();
-    }
-
-    /*
-     * Only save to use if called once it has been confirmed that the passed ItemStack is indeed a bundle.
-     * */
-    public static List<ItemStack> getBundleItems(ItemStack itemStack) {
-        return ((BundleMeta) itemStack.getItemMeta()).getItems();
     }
 }
